requirements.txt
mysql-connector-python==8.4.0
python-dotenv==1.0.1
pytest==8.3.3
pytest-cov==5.0.0
